requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
dj-database-url==1.2.0
Django==4.1.5
django-bootstrap4==22.3
django-mysql==4.8.0
django-stdimage==6.0.1
gunicorn==20.1.0
mysql==0.0.3
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2-binary==2.9.5
PyMySQL==1.0.2
python-decouple==3.6
python-dotenv==0.21.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
whitenoise==6.3.0
